% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.R
\name{add_function}
\alias{add_function}
\title{Split a string}
\usage{
add_function(x, y)
}
\arguments{
\item{x}{A numeric number}

\item{y}{A numeric number}
}
\value{
A numeric number.
}
\description{
Split a string
}
\examples{
add_function(1, 2)       # output: 3
add_function(1,3)          # output: 4
}
