% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regress.R
\name{regress}
\alias{regress}
\title{This to apply lm model to data}
\source{
The data comes from simulation.
}
\usage{
regress(dat)
}
\arguments{
\item{dat}{The input dataframe}
}
\value{
summary of lm model
}
\description{
This is short description
}
\details{
This is package
}
\references{
Nagarajan R, Scutari M, Lebre S (2013). "Bayesian Networks in R with Applications in Systems Biology". Springer.
}
\author{
author Fadhaa Ali
}
\concept{Linear Regression.}
